refactor(docs): follow the inverted specifier convention
docs chrome flips to bare base imports, radix-context components and the thread-list sample provider mark their imports explicitly (a base provider around a radix sidebar breaks react context at prerender), the manual tab display normalizes radix specifiers, demo zips flatten both trees, and the demo dependency list swaps the retired radix icons package for @base-ui/react.
