+++ apps/docs/app/(demos)/heat-graph/page.tsx
@@ -11,7 +11,7 @@ import {
 import Link from "next/link";
 import { cn } from "@/lib/utils";
 import { CopyCommandButton } from "@/components/home/copy-command-button";
-import { Button, buttonVariants } from "@/components/ui/button";
+import { Button, buttonVariants } from "@/components/ui/radix/button";
 import { HeatGraphDemo } from "./heat-graph-demo";
 
 const ANALYTICS_PAGE = "heat-graph" as const;
