+++ apps/docs/app/(demos)/react-o11y/page.tsx
@@ -12,7 +12,7 @@ import {
 import Link from "next/link";
 import { cn } from "@/lib/utils";
 import { CopyCommandButton } from "@/components/home/copy-command-button";
-import { Button, buttonVariants } from "@/components/ui/button";
+import { Button, buttonVariants } from "@/components/ui/radix/button";
 import { WaterfallSample } from "@/components/docs/samples/o11y/waterfall";
 import {
   CollapseSample,
