+++ apps/docs/components/docs/status-badge.tsx
@@ -5,7 +5,7 @@ import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
-} from "@/components/ui/tooltip";
+} from "@/components/ui/radix/tooltip";
 import { cn } from "@/lib/utils";
 
 const VARIANTS = {
