+++ apps/docs/components/xulux/shell/XuluxHistoryMenu.tsx
@@ -9,7 +9,7 @@ import {
   DropdownMenuItem,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from "@/components/ui-base/dropdown-menu";
+} from "@/components/ui/dropdown-menu";
 import { cn } from "@/lib/utils";
 import {
   useNormalizeInterruptedXuluxThreads,
