+++ apps/docs/content/docs/(reference)/api-reference/context.tsx
@@ -6,7 +6,7 @@ import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
-} from "@/components/ui/tooltip";
+} from "@/components/ui/radix/tooltip";
 import { Fragment } from "react";
 
 interface ComponentProps {
