+++ apps/docs/content/examples/perplexity.mdx
@@ -81,7 +81,7 @@ const Composer = ({ placeholder }) => (
 ### Functional Search & Model Dropdowns
 
 ```tsx
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui-base/dropdown-menu";
+import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu";
 
 const SEARCH_MODES = [
   { id: "search", name: "Search", description: "Fast answers to everyday questions", Icon: Search },
