+++ apps/docs/lib/xulux/demo-downloads/create-demo-zip.ts
@@ -93,7 +93,8 @@ function targetPathForSourceFile(sourceFile: string) {
   if (sourceFile.startsWith("packages/ui/src/")) {
     return sourceFile
       .replace(/^packages\/ui\/src\//, "")
-      .replace(/^components\/ui\//, "components/ui/")
+      .replace(/^components\/ui\/radix\//, "components/ui/")
+      .replace(/^components\/ui\/base\//, "components/ui/")
       .replace(/^components\/assistant-ui\//, "components/assistant-ui/")
       .replace(/^lib\//, "lib/");
   }
